// mcp/apps/mcp/api-schema.d.ts
export interface paths {
// ...
    "/auth/whoami": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /** @description Obtain information about the logged in user */
        get: operations["getAuthWhoami"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
    "/works/{id}": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get: operations["getWorkById"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
    "/search/{models}": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get: operations["getSearchWithModels"];
        put?: never;
        post: operations["postSearchWithModels"];
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/oai": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        get: operations["oaiGet"];
        put?: never;
        post: operations["oaiPost"];
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
}
export type webhooks = Record<string, never>;
export interface components {
// ...
    parameters: {
        /** @description Desired output format */
        as: "iiif" | "opensearch";
        /** @description Page number of results to retrieve */
        page: number;
        /** @description Maximum number of results per page */
        size: number;
        /** @description Comma-delimited list of fields to sort search results (e.g. "create_date:asc,modified_date:desc") */
        sort: string;
        /** @description Collection, FileSet, or Work ID */
        id: string;
        /** @description Desired aspect ratio */
        thumbnailAspect: "full" | "square";
        /** @description Size of largest dimension */
        thumbnailSize: number;
        /** @description Keyword query */
        query: string;
        /** @description Search token from previous search response */
        searchToken: string;
        /** @description Comma-delimited list of models to search */
        models: ("collections" | "file-sets" | "works")[];
    };
// ...
}
export type $defs = Record<string, never>;
export interface operations {
// ...
    postSearchWithModels: {
        parameters: {
            query?: {
                /** @description Page number of results to retrieve */
                page?: components["parameters"]["page"];
                /** @description Maximum number of results per page */
                size?: components["parameters"]["size"];
                /** @description Comma-delimited list of fields to sort search results (e.g. "create_date:asc,modified_date:desc") */
                sort?: components["parameters"]["sort"];
                /** @description Desired output format */
                as?: components["parameters"]["as"];
            };
            header?: never;
            path: {
                /** @description Comma-delimited list of models to search */
                models: components["parameters"]["models"];
            };
            cookie?: never;
        };
        requestBody?: components["requestBodies"]["search"];
        responses: {
            200: components["responses"]["SearchResponse"];
        };
    };
// ...
}